Refaktorering, la til requiresNoneOf for validering
